put common ops in a separate file
